refactor: controller 객체의 @Controller 어노테이션을 @RestController 어노테이션으로 변경
