+++ src/main/java/chess/controller/ChessGameController.java
@@ -11,8 +11,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
 
-@Controller
+@RestController
 @RequestMapping("/game")
 public class ChessGameController {
 
@@ -22,36 +23,26 @@ public class ChessGameController {
         this.gameService = gameService;
     }
 
-    @GetMapping("/{id}")
-    public String board() {
-        return "board";
-    }
-
-    @ResponseBody
     @PostMapping("/{id}/start")
     public BoardDto start(@PathVariable("id") String id) {
         return gameService.start(id);
     }
 
-    @ResponseBody
     @PostMapping("/{id}/end")
     public BoardDto end(@PathVariable("id") String id) {
         return gameService.end(id);
     }
 
-    @ResponseBody
     @GetMapping("/{id}/status")
     public ResultDto status(@PathVariable("id") String id) {
         return gameService.status(id);
     }
 
-    @ResponseBody
     @GetMapping("/{id}/load")
     public BoardDto load(@PathVariable("id") String id) {
         return gameService.load(id);
     }
 
-    @ResponseBody
     @PostMapping("/{id}/move")
     public BoardDto move(@PathVariable("id") String id, @RequestBody MoveDto moveDto) {
         return gameService.move(id, moveDto);
